+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # Change Log
 
+### 0.0.3
+
+- Adds stashing switches
+- Adds stash "keep index"